requirements.txt
google-genai==1.70.0
python-dotenv==1.2.2
pyvts==0.3.3
aiofiles==25.1.0
websockets==16.0
pydantic==2.12.5
elevenlabs==2.41.0
SpeechRecognition==3.14.3
# Required for xAI (Grok) Support
openai
xai-sdk
# Optional: May fail on some Python versions
PyAudio==0.2.14